@@ -13,7 +13,10 @@
     "build": "vite build",
     "build:check": "vue-tsc --noEmit && vite build",
     "preview": "vite preview",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx --fix --ignore-path .gitignore",
+    "lint": "node run-eslint.js",
+    "lint:fix": "eslint --fix src/**/*.{vue,js,jsx,ts,tsx}",
+    "build:nolint": "SKIP_LINT=true vite build",
+    "build:warn": "ESLINT_WARNINGS_ONLY=true vite build",
     "typecheck": "vue-tsc --noEmit",
     "deploy:local": "node scripts/deploy.js",
     "deploy": "npm run build && gh-pages -d dist",
@@ -38,13 +41,16 @@
   "devDependencies": {
     "@inspira-ui/plugins": "0.0.1",
     "@types/node": "^20.11.19",
+    "@typescript-eslint/eslint-plugin": "7.18.0",
+    "@typescript-eslint/parser": "7.18.0",
     "@vitejs/plugin-vue": "^4.4.0",
     "@vue/tsconfig": "^0.5.1",
     "autoprefixer": "^10.4.16",
     "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "concurrently": "9.1.2",
+    "eslint": "8.57.0",
+    "eslint-plugin-vue": "9.33.0",
     "gh-pages": "6.3.0",
     "postcss": "^8.4.31",
     "tailwind-merge": "3.0.2",
@@ -52,6 +58,8 @@
     "tailwindcss-animate": "1.0.7",
     "typescript": "~5.2.0",
     "vite": "5.2.0",
+    "vite-plugin-eslint": "1.8.1",
+    "vue-eslint-parser": "9.4.3",
     "vue-tsc": "^1.8.25"
   }
 }
