@@ -3,23 +3,32 @@
     <AppHeader />
     
     <main class="container mx-auto py-6 px-4">
-      <h1 class="text-2xl font-bold mb-4">Welcome to DreamMall Matching</h1>
+      <h1 class="text-2xl font-bold mb-4">
+        Welcome to DreamMall Matching
+      </h1>
       
       <div class="grid grid-cols-1 md:grid-cols-2 gap-6">
         <div class="bg-white p-6 rounded-lg shadow-md">
-          <h2 class="text-xl font-semibold mb-3">User Profiles</h2>
-          <p class="mb-4">View and manage user profiles</p>
+          <h2 class="text-xl font-semibold mb-3">
+            User Profiles
+          </h2>
+          <p class="mb-4">
+            View and manage user profiles
+          </p>
           
           <div class="mb-4">
             <input
-              type="text"
               v-model="searchTerm"
+              type="text"
               placeholder="Search users..."
               class="w-full px-3 py-2 border border-gray-300 rounded-md"
-            />
+            >
           </div>
           
-          <div v-if="filteredUsers.length > 0" class="space-y-2 max-h-96 overflow-y-auto">
+          <div
+            v-if="filteredUsers.length > 0"
+            class="space-y-2 max-h-96 overflow-y-auto"
+          >
             <div 
               v-for="user in filteredUsers" 
               :key="user.id" 
@@ -27,8 +36,13 @@
             >
               <div>
                 <span class="font-medium">{{ user.name }}</span>
-                <p class="text-gray-500 text-sm">{{ user.email }}</p>
-                <div v-if="user.skills && user.skills.length" class="mt-1">
+                <p class="text-gray-500 text-sm">
+                  {{ user.email }}
+                </p>
+                <div
+                  v-if="user.skills && user.skills.length"
+                  class="mt-1"
+                >
                   <span 
                     v-for="skill in user.skills.slice(0, 3)" 
                     :key="skill"
@@ -36,7 +50,10 @@
                   >
                     {{ skill }}
                   </span>
-                  <span v-if="user.skills.length > 3" class="text-xs text-gray-500">
+                  <span
+                    v-if="user.skills.length > 3"
+                    class="text-xs text-gray-500"
+                  >
                     +{{ user.skills.length - 3 }} more
                   </span>
                 </div>
@@ -51,29 +68,52 @@
               </div>
             </div>
           </div>
-          <p v-else-if="userStore.isLoading" class="text-center py-4">Loading users...</p>
-          <p v-else class="text-center py-4">No users available</p>
+          <p
+            v-else-if="userStore.isLoading"
+            class="text-center py-4"
+          >
+            Loading users...
+          </p>
+          <p
+            v-else
+            class="text-center py-4"
+          >
+            No users available
+          </p>
         </div>
         
         <div class="bg-white p-6 rounded-lg shadow-md">
-          <h2 class="text-xl font-semibold mb-3">Import Profiles</h2>
-          <p class="mb-4">Upload a JSON file to import user profiles</p>
+          <h2 class="text-xl font-semibold mb-3">
+            Import Profiles
+          </h2>
+          <p class="mb-4">
+            Upload a JSON file to import user profiles
+          </p>
           <label class="block mb-4">
             <span class="sr-only">Choose file</span>
-            <input type="file" accept=".json" 
-                  class="block w-full text-sm text-gray-500
+            <input
+              type="file"
+              accept=".json" 
+              class="block w-full text-sm text-gray-500
                   file:mr-4 file:py-2 file:px-4
                   file:rounded-md file:border-0
                   file:text-sm file:font-semibold
                   file:bg-blue-50 file:text-blue-700
                   hover:file:bg-blue-100"
-                  @change="importProfilesFromFile"/>
+              @change="importProfilesFromFile"
+            >
           </label>
           
           <div class="border-t pt-4 mt-4">
-            <h3 class="font-medium mb-2">Default Test User</h3>
-            <p class="text-sm text-gray-600 mb-2">Email: test@example.com</p>
-            <p class="text-sm text-gray-600">Password: test123</p>
+            <h3 class="font-medium mb-2">
+              Default Test User
+            </h3>
+            <p class="text-sm text-gray-600 mb-2">
+              Email: test@example.com
+            </p>
+            <p class="text-sm text-gray-600">
+              Password: test123
+            </p>
           </div>
         </div>
       </div>
