@@ -1,10 +1,16 @@
 <template>
   <div class="container mx-auto p-4">
-    <h1 class="text-3xl font-bold mb-4">DreamMall Matching System</h1>
-    <p class="mb-4">Welcome to the DreamMall Matching System! This application helps connect people based on shared interests and complementary skills.</p>
+    <h1 class="text-3xl font-bold mb-4">
+      DreamMall Matching System
+    </h1>
+    <p class="mb-4">
+      Welcome to the DreamMall Matching System! This application helps connect people based on shared interests and complementary skills.
+    </p>
     
     <div class="bg-blue-100 p-4 rounded-lg mb-4">
-      <h2 class="text-xl font-bold mb-2">Getting Started</h2>
+      <h2 class="text-xl font-bold mb-2">
+        Getting Started
+      </h2>
       <p>Explore the different features using the navigation above.</p>
     </div>
   </div>
