@@ -1,8 +1,17 @@
 <template>
   <div class="container mx-auto p-4 text-center">
-    <h1 class="text-4xl font-bold mb-4">404</h1>
-    <p class="text-xl mb-4">Page not found</p>
-    <router-link to="/" class="text-blue-500 hover:underline">Return to Home</router-link>
+    <h1 class="text-4xl font-bold mb-4">
+      404
+    </h1>
+    <p class="text-xl mb-4">
+      Page not found
+    </p>
+    <router-link
+      to="/"
+      class="text-blue-500 hover:underline"
+    >
+      Return to Home
+    </router-link>
   </div>
 </template>
 
